Replaces the javascript padding with a css transformX
We were starting the root node (li.node) at -10px since we don't show
the toggler on the root node, so we don't need the space. Then each
calculation was incorporating this 10px padding. Rather than adding that
complexity to the JS, I'm shifting the ul -10px using a transformX so
it's done once.