Stores internal _nodeData as sparse array.
The key is the node id and the value is the node. When we remove nodes,
this will be easier/faster to sync _nodeData. Additionally, getting a
node's parent will be much faster as we're building the tree.

An object could also be used, and might be faster, but I don't think
this is something that will be noticable. The array gives us some extra
convenience from a code standpoint, and we can always optimize laster.

But it won't be a problem for <100K nodes I'm sure.