Adds the ability to set a visible prop on a node.
This can be used to make sure a node is not visible. The node will be
stored on the parent's _invisibleNodes property, so we can continue to
swap between children and _children w/ wreckless abandon.

A patch(..) can change the visible property, which will move the node
back to the parent's proper array or remove it from that array and then
add it to _invisibleNodes.

Fixes #75