Adds a forest tree.
If there are multiple nodes that are missing a parent, the tree turns
into multiple subtrees. Meaning each node that is missing a parent
becomes a root.

The tree.root becomes an array of all roots, and we build nodes by
reducing the root array into a list of all nodes.

Fixes #77