Rewrites the way we control children.
Instead of flopping between .children and ._children, with the added
touch of throwing `visible: false` nodes in _.invisibleNodes, we throwt
all children in an array ._allChildren on the parent. Then the d3 tree
children accessor function is set to control what nodes will be part of
the real .children (which is what's visible to the end user in the DOM).

Fixes #146