Fixes _transitionWrap logic to account for visible descendants.
Previously if you had a tree that looked like this:

```
Grandparent
  Parent
    Child
      Newborn
      Newborn
      Newborn
      Newborn
      Newborn
      Newborn
      Newborn
```
And you collapsed the parent node, and then expanded it, we would
determine if transitions should be turned of by looking at the parent
children count. But in this case that value would be 1, but when the
tree expands, we'd show all Newborn node's as well. This now reduces
each child with all its visible descendants to get a true expansition
count.

Fixes #229