Adds dnd tree autoscrolling.
This also fixes the way we determine the lower and upper bounds for a
traveling node. Previously it was tied to tree.node(), but that doesn't
work since a node w/ 1000 children would have a last node w/ a very
large y value, meaning the traveling node would be way below the tree.
Same thing would apply if the tree was scrolled to the bottom, and the
user wanted to move a node up.

The autoscroller starts an interval that will keep scrolling a tree,
even if the mouse doesn't move.

Fixes #234