Improves dnd _move logic.
This ditches an approach I used previously by figuring out the index of the
traveling node within `tree.node`. This works if all the elements are in
the dom, but fails if we want to hide some nodes outside of the viewport
(for performance purposes.)

I tried multi approaches for figuring out where we're dropping the node.
I tried to ask the dom which element is at a location. This actually
works really well. I would hide the traveling node, get the dom element
underneath it, which is a `li.node`, and then show the traveling node
again. However this fails w/ transitions if the user is moving their
mouse quickly, so we have do things with math + data.

Generally, this cleans up a lot of the move code, which should help
figure out a solution to #174 as well.