Rewrites (read: cleanups) dnd.
There were quite a few things I found that were wrong.

1. In a forest tree, if a node was at root, then moved to a child of a
root sibling, then moved back out to a root node, it still had a parent
node set from when it was indented. This wasn't immediately obvious in
the UI, but the data was bad.

2. If you wanted to prevent nodes from being embedded, the default
`embed` logic wasn't used, it wouldn't correctly figure out the index of
the new node's location. Basically there was some bad +1 or -1 logic for
determing an index.

3. General cleanup of dnd so it's not as poorly written. This is much
more modular and readable, but still somewhat confusing.