Optimizes large search results.
This commit introduces a new function `._join` which binds data to the
tree selection. Previously all that magic happened in `_rebind`. Now
`_rebind` will determine the data model in the tree, by default, and
_join is where the data is bound to the dom selection.

This is used by search results, so we can reuse the logic that
determines is a tree should be optimized.

Sadly, search results now require us to store the search results data.
This is because of adjusting the viewport on scrolling. If we didn't do
this, the next time a user would scroll when viewing search results, it
would rebind to the original data in the tree, which is wrong.

Fixes #250