Adds the ability to override root node height.
By default, the root node will have a height equal to the default height
(i.e. 36px). Some trees may want a different (larger) style for the root
node. The tree receives an option `rootHeight` which will be used for
the root node's height. If a tree has a root node height that is larger
than the default node height, the tree has a class of `detached-root`
applied.

Fixes #275