Rewrites update code so it utilizes `contents`.
In the previous implementation, we allowed a `contents` options override
that was used just on `li.node` enter code. But this was broken for two
reasons.

1. The update code (ie. update.js) worked on a node structure that
banked on contents being the default.

2. All tree code for setting indentation levels used a selector for
`node-contents`, but if a caller overriding options.contents didn't set
that class, indentations would break.

This makes the overridable contents be a proper d3 selection, so calling
code can hook into d3 data joining (enter/update/exit).

Fixes #324