WIP Sets a node's z-index to be based on its depth.
Now when a node is collapsed to its parent (or some ancestor), it will
be stacked below it. To prevent the lower stacked nodes from bleeding,
the .node needs to have a defined background-color, so that color is now
set in the .node, as well as the .tree-container (so extra space in the
tree has the same color).

e.g.
Root: (z-index 0)
  P1 (z-index -1)
    O1 (z-index -2)
    O2 (z-index -3)
      M1 (z-index -4)
  P2 (z-index -1)

Fixes #338