Sets important when we fade out opacity.
Some of the css selectors have an `!important` priority set, which isn't
really ideal, but we set those priorities because we are constantly
changing the inline opacity when we transition nodes in/out.

When a tree was in DND node, the important priority on a node's opacity
was preventing our inline fade out opacity from going to 0, so editable
nodes would never truly fade out.