Force overflow:visible on li.nodes
This seems to be necessary because of a firefox rendering bug. The
problem can be demonstrated here:

http://codepen.io/anon/pen/VmRROj

When the li.node has overflow-x and overflow-y `hidden`, firefox will
break on `li.node .node-contents .label` innerHTML changes. After a node
moves, if you try to change its label, the label disappears from the
browser, even though inspecting them dom shows the label is correct.
Removing these properties altogether work in the tree project, but still
show bugs in SB3 (driving force behind this project), but forcing
overflow to visible seems to hide the problem in FF. Seems like one of
those "This fixes it, but shouldn't" commits.

It's also scary removing the hidden property, but it seems like we're
okay without out, since we're using label-mask.

Fixes #344
Also Fixes Sb3 issue #9216