Emits `rendered` events after we're done transitioning the tree.
Previously we listened to `transitionend` on the node selection. This
would fire multiple times for each node that was moving, since
transitionend fires for each property that's changing. For a node in the
tree, that could be its `opacity`, `transform left`, and/or `transform
top`. This approach uses the timeout setting we for transitions, so the
callback is fired just once. When the nodes are done moving/updating, we
remove the `transitions` class on the tree, and fire a tree
`rendered` event for any interested parties.