Fires `selected` events when a node is selected and visible in the tree.
This fires after the `select` event on a node. `select` event fires
immediately, whereas `selected` will fire after that node has been
selected and is visible in the tree, meaning transitions have finished
and the tree has been scrolled so the node is now in the viewport.

Drops the `rendered` event we were previously using. It's too buggy
since some updates to the tree won't end up going into the
`transitionWrap` callback.