Remove `transition: none` for edit mode transition nodes.
In
https://github.com/SpiderStrategies/kalpa-tree/commit/579c3c8192fe009122d1a7e3985a1f1d2823e254
we added this transition: none; to fix an issue with IE11. Derek and I
remember testing this change with @scottoreilly, but apparently we
missed something. When I go back to this commit, it's what introduced
the new bug in Sb3 (#12556). It doesn't appear like this `transition:
none;` is needed.

Other issues for history:
https://github.com/SpiderStrategies/Scoreboard/issues/8950
https://github.com/SpiderStrategies/kalpa-tree/pull/362
https://github.com/SpiderStrategies/kalpa-tree/pull/363

Fixes https://github.com/SpiderStrategies/Scoreboard/issues/12556